+++ sites/geohub/src/components/controls/RasterIntervalsLegend.svelte
@@ -1,30 +1,29 @@
 <script lang="ts">
   import { onMount } from 'svelte'
-  import { PUBLIC_TITILER_ENDPOINT } from '$lib/variables/public'
   import {
     ClassificationMethodNames,
     ClassificationMethodTypes,
     COLOR_CLASS_COUNT_MAXIMUM,
     COLOR_CLASS_COUNT_MINIMUM,
   } from '$lib/constants'
-  import { cloneDeep, debounce } from 'lodash-es'
+  import { cloneDeep } from 'lodash-es'
   import {
-    fetchUrl,
     generateColorMap,
     getActiveBandIndex,
     getLayerStyle,
-    getValueFromRasterTileUrl,
     updateParamsInURL,
     getLayerSourceUrl,
     getMaxValueOfCharsInIntervals,
+    remapInputValue,
   } from '$lib/helper'
   import NumberInput from '$components/controls/NumberInput.svelte'
   import IntervalsLegendColorMapRow from '$components/controls/IntervalsLegendColorMapRow.svelte'
-  import type { IntervalLegendColorMapRow, Layer, RasterLayerStats, RasterTileMetadata } from '$lib/types'
-  import { layerList, map } from '$stores'
+  import type { IntervalLegendColorMapRow, Layer } from '$lib/types'
+  import { map } from '$stores'
   import { updateIntervalValues } from '$lib/helper/updateIntervalValues'
+  import ColorMapPicker from './ColorMapPicker.svelte'
 
-  export let colorPickerVisibleIndex: number
+  //console.clear()
   export let layerConfig: Layer
   export let numberOfClasses: number
   export let colorClassCountMax = COLOR_CLASS_COUNT_MAXIMUM
@@ -32,32 +31,44 @@
   export let colorMapName: string
   export let classificationMethod: ClassificationMethodTypes
   export let colorMapRows: Array<IntervalLegendColorMapRow>
+  // this var is necessary to maintain the state of teh colormap when switching the legend.
+  // and it should be set by the bool flags that control the colormap picker visibility from parent container
+  export let generateCmap: boolean
 
-  $: colorMapName, colorManNameChanged()
+  $: colorMapName, colorMapNameChanged()
 
-  const colorManNameChanged = () => {
-    //console.log('RIL', colorMapName, colorMapRows.length)
-    //if (colorMapRows.length > 0) return
+  const colorMapNameChanged = () => {
+    // this also happens on init/mounting. because of this it is not posisble to know when the colormap has been changed
+    // or when the component is being created. On demand (event based approach) might be better
 
-    if (colorMapName) reclassifyImage() // not right
+    // the map is updated whene the colormap is chenged intentionally or whne it is initialized first time
+    if ((colorMapName && generateCmap) || (colorMapName && colorMapRows.length == 0)) {
+      reclassifyImage()
+    } // not right
+    else {
+      handleParamsUpdate()
+    }
   }
 
   let { info }: Layer = layerConfig
   const bandIndex = getActiveBandIndex(info)
-
   let layerMax
   let layerMin
+  let layerMean
   let rowWidth: number
   if ('stats' in info) {
     const band = Object.keys(info.stats)[bandIndex]
+
     layerMin = Number(info.stats[band].min)
     layerMax = Number(info.stats[band].max)
+    layerMean = Number(info.stats[band].mean)
   } else {
     const [band, bandMetaStats] = info['band_metadata'][bandIndex]
     layerMin = Number(bandMetaStats['STATISTICS_MINIMUM'])
     layerMax = Number(bandMetaStats['STATISTICS_MAXIMUM'])
+    layerMean = Number(bandMetaStats['STATISTICS_MEAN'])
   }
-
+  const layerMeanToMax = layerMean / layerMax
   let percentile98: number = info.stats[Object.keys(info.stats)[bandIndex]]['percentile_98']
   let classificationMethods = [
     { name: ClassificationMethodNames.NATURAL_BREAK, code: ClassificationMethodTypes.NATURAL_BREAK },
@@ -66,8 +77,14 @@
     { name: ClassificationMethodNames.LOGARITHMIC, code: ClassificationMethodTypes.LOGARITHMIC },
   ]
 
+  if (layerMeanToMax >= -0.5 && layerMeanToMax <= 0.5) classificationMethod = ClassificationMethodTypes.LOGARITHMIC
+  if ((layerMeanToMax > -5 && layerMeanToMax < -0.5) || (layerMeanToMax > 0.5 && layerMeanToMax < 5))
+    classificationMethod = ClassificationMethodTypes.NATURAL_BREAK
+  if (layerMeanToMax <= -5 && layerMeanToMax >= 5) classificationMethod = ClassificationMethodTypes.EQUIDISTANT
+
   onMount(async () => {
-    const rasterInfo = info as RasterTileMetadata
+    //const rasterInfo = info as RasterTileMetadata
+
     // if (!rasterInfo?.isMosaicJson) {
     //   const layerUrl = getLayerSourceUrl($map, layerConfig.id) as string
 
@@ -77,46 +94,61 @@
     //   const layerStats: RasterLayerStats = await fetchUrl(statsURL)
     //   info = { ...info, stats: layerStats }
     // }
-    const band = info.active_band_no
-    percentile98 = info.stats[band]['percentile_98']
-    const skewness = 3 * ((info.stats[band].mean - info.stats[band].median) / info.stats[band].std)
-    //TODO discuss with Joseph
-    // if (classificationMethod === ClassificationMethodTypes.LOGARITHMIC) {
-    //   if (skewness > 1 && skewness > -1) {
-    //     // Layer isn't higly skewed.
+    // const band = info?.active_band_no
+    // percentile98 = info?.stats[band]['percentile_98']
 
-    //     classificationMethod = ClassificationMethodTypes.EQUIDISTANT // Default classification method
-    //   } else {
-    //     classificationMethod = ClassificationMethodTypes.LOGARITHMIC
-    //   }
-    // }
+    // //skewness based
+    // const skewness = 3 * ((info.stats[band].mean - info.stats[band].median) / info.stats[band].std)
+    // if (skewness>-.5 && skewness < .5) classificationMethod = ClassificationMethodTypes.EQUIDISTANT
+    // if (skewness>-1 && skewness<-.5 || skewness>.5 && skewness < 1) classificationMethod = ClassificationMethodTypes.NATURAL_BREAK
+    // if (skewness<-1 || skewness> 1) classificationMethod = ClassificationMethodTypes.LOGARITHMIC
 
-    layerConfig = { ...layerConfig, info: info }
-    const layers = $layerList.map((layer) => {
-      return layerConfig.id !== layer.id ? layer : layerConfig
-    })
-    layerList.set([...layers])
-    getColorMapRows()
-    colorMapRows.length > 0 ? null : reclassifyImage()
+    //based on mean/max ratio
+
+    // if (layerMeanToMax >= -0.5 && layerMeanToMax <= 0.5) classificationMethod = ClassificationMethodTypes.LOGARITHMIC
+    // if ((layerMeanToMax > -5 && layerMeanToMax < -0.5) || (layerMeanToMax > 0.5 && layerMeanToMax < 5))
+    //   classificationMethod = ClassificationMethodTypes.NATURAL_BREAK
+    // if (layerMeanToMax <= -5 && layerMeanToMax >= 5) classificationMethod = ClassificationMethodTypes.EQUIDISTANT
+
+    // layerConfig = { ...layerConfig, info: info }
+    // const layers = $layerList.map((layer) => {
+    //   return layerConfig.id !== layer.id ? layer : layerConfig
+    // })
+    // layerList.set([...layers])
+
+    // this fucntion is useless, really, at the time when on mount is executed the map has not been yet updated and
+    // the colormap object does not exist in the layer's url. A forced sync with the map might be better
+    // and could work. However, with the introduction of colorMapRows in te state management
+    // this funcntion becomes obsolete. and should be removed
+
+    if (colorMapRows.length == 0) {
+      reclassifyImage()
+    }
   })
 
-  const getColorMapRows = () => {
-    const colormap: number[][][] = getValueFromRasterTileUrl($map, layerConfig.id, 'colormap') as number[][][]
-    colorMapRows = []
-    if (colormap && colormap.length > 0) {
-      colormap.forEach((row: number[][], index: number) => {
-        const values = row[0]
-        const color = row[1]
-        colorMapRows.push({
-          color: color,
-          index: index,
-          start: values[0],
-          end: values[1],
-        })
-      })
-    }
-    numberOfClasses = colorMapRows.length === 0 ? numberOfClasses : colorMapRows.length
-  }
+  // const getColorMapRows = () => {
+  //   const layerUrl = getLayerSourceUrl($map, layerConfig.id) as string
+  //   const layerURL = new URL(layerUrl)
+  //   console.log(layerUrl)
+
+  //   const colormap: number[][][] = getValueFromRasterTileUrl($map, layerConfig.id, 'colormap') as number[][][]
+
+  //   console.log(colormap)
+  //   colorMapRows = []
+  //   if (colormap && colormap.length > 0) {
+  //     colormap.forEach((row: number[][], index: number) => {
+  //       const values = row[0]
+  //       const color = row[1]
+  //       colorMapRows.push({
+  //         color: color,
+  //         index: index,
+  //         start: values[0],
+  //         end: values[1],
+  //       })
+  //     })
+  //   }
+  //   numberOfClasses = colorMapRows.length === 0 ? numberOfClasses : colorMapRows.length
+  // }
 
   const reclassifyImage = (e?: CustomEvent) => {
     let isClassificationMethodEdited = false
@@ -147,9 +179,20 @@
     })
   }
 
+  // it is very interesting that without debounce it does NOW properly
   // encode colormap and update url parameters
-  const handleParamsUpdate = debounce(() => {
-    const encodeColorMapRows = JSON.stringify(colorMapRows.map((row) => [[row.start, row.end], row.color]))
+  const handleParamsUpdate = () => {
+    const encodeColorMapRows = JSON.stringify(
+      colorMapRows.map((row) => {
+        if (row.color[3] === 255) {
+          return [[row.start, row.end], row.color]
+        } else {
+          const a = remapInputValue(row.color[3], 0, 1, 0, 255)
+          const color = [row.color[0], row.color[1], row.color[2], Math.floor(a)]
+          return [[row.start, row.end], color]
+        }
+      }),
+    )
     const layerUrl = getLayerSourceUrl($map, layerConfig.id) as string
     if (!(layerUrl && layerUrl.length > 0)) return
     const layerURL = new URL(layerUrl)
@@ -158,7 +201,7 @@
     const updatedParams = Object.assign({ colormap: encodeColorMapRows })
     const layerStyle = getLayerStyle($map, layerConfig.id)
     updateParamsInURL(layerStyle, layerURL, updatedParams)
-  }, 500)
+  }
 
   const handleIncrementDecrementClasses = (e: CustomEvent) => {
     numberOfClasses = e.detail.value
@@ -168,9 +211,6 @@
     reclassifyImage()
   }
 
-  const handleColorPickerClick = (event: CustomEvent) => {
-    colorPickerVisibleIndex = event.detail.index
-  }
   const handleChangeIntervalValues = (event: CustomEvent) => {
     colorMapRows = updateIntervalValues(event, colorMapRows)
     rowWidth = getMaxValueOfCharsInIntervals(colorMapRows)
@@ -182,19 +222,15 @@
   class="intervals-view-container"
   data-testid="intervals-view-container">
   <!-- svelte-ignore a11y-click-events-have-key-events -->
-  <div
-    class="columns is-gapless controls"
-    on:click={() => (colorPickerVisibleIndex = -1)}>
-    <div class="column classification">
-      <div class="has-text-centered pb-2">Classification</div>
-      <div
-        class="select is-flex is-justify-content-center"
-        style="height: 30px; width: fit-content">
+  <div class="legend-controls">
+    <div class="classification field pr-2">
+      <!-- svelte-ignore a11y-label-has-associated-control -->
+      <label class="label has-text-centered">Classification</label>
+      <div class="control">
         <select
           bind:value={classificationMethod}
           on:change={(e) => reclassifyImage(e)}
           style="width: 114px;"
-          alt="Classification Methods"
           title="Classification Methods">
           {#each classificationMethods as classificationMethod}
             <option
@@ -205,13 +241,25 @@
       </div>
     </div>
 
-    <div class="column number-classes">
-      <div class="has-text-centered">Number of Classes</div>
-      <NumberInput
-        bind:value={numberOfClasses}
-        bind:minValue={colorClassCountMin}
-        bind:maxValue={colorClassCountMax}
-        on:change={handleIncrementDecrementClasses} />
+    <div class="number-classes field pr-2">
+      <!-- svelte-ignore a11y-label-has-associated-control -->
+      <label class="label has-text-centered">Number of Classes</label>
+      <div class="control">
+        <NumberInput
+          bind:value={numberOfClasses}
+          bind:minValue={colorClassCountMin}
+          bind:maxValue={colorClassCountMax}
+          on:change={handleIncrementDecrementClasses} />
+      </div>
+    </div>
+
+    <div class="colormap-picker">
+      <ColorMapPicker
+        bind:colorMapName
+        on:colorMapChanged={() => {
+          colorMapNameChanged()
+          reclassifyImage()
+        }} />
     </div>
   </div>
   <div class="is-divider separator mb-4" />
@@ -221,9 +269,6 @@
       bind:colorMapRow
       bind:colorMapName
       bind:rowWidth
-      {colorPickerVisibleIndex}
-      on:clickColorPicker={handleColorPickerClick}
-      on:closeColorPicker={() => (colorPickerVisibleIndex = -1)}
       on:changeColorMap={handleParamsUpdate}
       on:changeIntervalValues={handleChangeIntervalValues} />
   {/each}
@@ -237,5 +282,19 @@
   }
   :global(.select:not(.is-multiple):not(.is-loading)::after) {
     border-color: #ff0000;
+  }
+
+  .legend-controls {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+
+    .number-classes {
+      margin: 0 auto;
+    }
+
+    .colormap-picker {
+      margin-left: auto;
+    }
   }
 </style>