+++ sites/geohub/src/components/controls/vector-styles/PropertySelect.svelte
@@ -7,8 +7,7 @@
   export let layer
   export let propertySelectValue
   export let showEmptyFields = false
-  export let showOnlyNumberFields = false
-  export let inLegend = false
+  export let inLegend
   export let emptyFieldLabel = 'No Label'
 
   let propertySelectOptions: string[]
@@ -16,23 +15,15 @@
   const dispatch = createEventDispatcher()
 
   onMount(() => {
-    inLegend && !propertySelectOptions ? setPropertyList() : null
-    !inLegend ? setPropertyList() : null
+    setPropertyList()
   })
 
   function setPropertyList() {
-    const vectorLayerMeta = getLayerProperties($map, layer, showOnlyNumberFields)
+    const vectorLayerMeta = getLayerProperties($map, layer, inLegend)
     propertySelectOptions = Object.keys(vectorLayerMeta.fields)
     if (showEmptyFields === true) {
       propertySelectOptions = ['', ...propertySelectOptions]
     }
-    propertySelectValue = setDefaultProperty(propertySelectOptions)
-    propertyChanged()
-  }
-
-  export let setDefaultProperty = (selectOptions: string[]) => {
-    if (selectOptions.length === 0) return ''
-    return selectOptions[0]
   }
 
   const propertyChanged = () => {
@@ -40,11 +31,8 @@
       prop: propertySelectValue,
     })
   }
-
-  $: propertySelectValue, propertyChanged()
 </script>
 
-<!--<div style="width: 100%; display: flex; align-items: center; justify-content: left; margin: auto">-->
 <div class="control has-icons-left">
   <div
     style="margin-right: 2%"
@@ -53,6 +41,7 @@
       style="width: 100%"
       class="is-normal"
       bind:value={propertySelectValue}
+      on:change={propertyChanged}
       alt="Property Options"
       title="Property Options">
       {#if propertySelectOptions}
@@ -73,6 +62,5 @@
   </span>
 </div>
 
-<!--</div>-->
 <style lang="scss">
 </style>