+++ apps/electricity-dashboard/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @undp-data/electricity-dashboard
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [5c98d25]
+  - @undp-data/svelte-undp-design@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [310179a]
+  - @undp-data/svelte-undp-design@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes